object: write objects atomically

+++ object.c
@@ -142,10 +142,38 @@ int object_write(ObjectType type, const void *data, size_t len, ObjectID *id_out
         return -1;
     }
 
-    close(fd);
-    unlink(tmp_path);
+    size_t written_total = 0;
+    while (written_total < total_len) {
+        ssize_t n = write(fd, object + written_total, total_len - written_total);
+        if (n <= 0) {
+            close(fd);
+            unlink(tmp_path);
+            free(object);
+            return -1;
+        }
+        written_total += (size_t)n;
+    }
+
+    if (fsync(fd) != 0 || close(fd) != 0) {
+        unlink(tmp_path);
+        free(object);
+        return -1;
+    }
+
+    if (rename(tmp_path, final_path) != 0) {
+        unlink(tmp_path);
+        free(object);
+        return -1;
+    }
+
+    int dirfd = open(shard_dir, O_RDONLY);
+    if (dirfd >= 0) {
+        fsync(dirfd);
+        close(dirfd);
+    }
+
     free(object);
-    return -1;
+    return 0;
 }
 
 // Read an object from the store.
